Upload #3
The game is almost finished...
The only thing left is to make a player take a card from the deck only after the sequence of moves were made, not immediately after making a move. Also, I need to try and make more readable UI. I guess I am finished with all the rules of the game.